+++ src/Mod/Points/App/Points.cpp
@@ -235,7 +235,7 @@ void PointKernel::save(std::ostream& out) const
 
 void PointKernel::getPoints(std::vector<Base::Vector3d> &Points,
                             std::vector<Base::Vector3d> &/*Normals*/,
-                            float /*Accuracy*/, uint16_t /*flags*/) const
+                            double /*Accuracy*/, uint16_t /*flags*/) const
 {
     unsigned long ctpoints = _Points.size();
     Points.reserve(ctpoints);
